Refactor import paths for CartContext and CartSheetProvider to use the lib directory
