+++ src/components/Header.tsx
@@ -3,8 +3,8 @@ import React, { useEffect, useState } from "react";
 import HeaderLogo from "@/components/HeaderLogo";
 import Navigation from "@/components/Navigation";
 import SearchBar from "@/components/SearchBar";
-import { useCartSheet } from "@/hooks/use-CartSheetProvider";
-import { useCart } from "@/hooks/use-CartContext";
+import { useCartSheet } from "@/lib/hooks/use-CartSheetProvider";
+import { useCart } from "@/lib/hooks/use-CartContext";
 import { Button } from "@/components/ui/button";
 import HeaderBox from "./HeaderBox";
 import { getLoggedInUser } from "@/actions/user.actions";
