+++ src/lib/hooks/useFetchProducts.ts
